requirements.txt
alpaca-py
yfinance
pandas
numpy
matplotlib
python-dotenv
lxml
beautifulsoup4
